Добавлен основной функционал сервиса.

+++ main/src/main/java/ru/practicum/controller/CompilationController.java
@@ -3,11 +3,9 @@ package ru.practicum.controller;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
-import org.springframework.stereotype.Controller;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import ru.practicum.dto.CompilationDto;
-import ru.practicum.model.Compilation;
 import ru.practicum.service.CompilationService;
 
 import javax.validation.constraints.Positive;
@@ -27,8 +25,8 @@ public class CompilationController {
     @ResponseStatus(HttpStatus.OK)
     public List<CompilationDto> getFilteredCompilations(
                 @RequestParam(required = false) Boolean pinned,
-                @PositiveOrZero @RequestParam(name = "from", defaultValue = "0") Integer from,
-                @Positive @RequestParam(name = "size", defaultValue = "10") Integer size) {
+                @PositiveOrZero @RequestParam(name = "from", defaultValue = "0") Long from,
+                @Positive @RequestParam(name = "size", defaultValue = "10") Long size) {
         List<CompilationDto> response = compilationService.getFilteredCompilations(pinned, from, size);
         log.info("Получен список подборок");
         return response;
